Fixed getTime
the atoi function from the standard c library does not work with the strings returned from the string object for whatever reason, despite the function returning a char. fixed the problem by writing a quick little ascii converter and changing the getTime function to work with the new ascToInt function. Also changed the variables in getTime to be two different strings for the date and the time. This was done because the function hangs on the second call to date after the first data has been processed. Getting both data at the start and then processing it fixes it. Also changed the function so that it does not use break; and instead uses a boolean to control the in/out of the loop. Consider using flush().