Added gobetwino and data compaction
Gobetwino functionality is currently commented out for just the serial tests. The data compaction follows that there is 4 decimal bits in the data incoming. This is the bit shifted over by 28,14, and 0 respectively to luminosity, humidity and temperature. Then using an appropriate bit shift to a 14 bit 1's variable, it is captured using a bitwise and.